refactor project creation flow
