+++ packages/skaff-lib/src/core/diffing/TemporaryProjectFactory.ts
@@ -4,7 +4,7 @@ import { ProjectSettings } from "@timonteutelink/template-types-lib";
 
 import { Result } from "../../lib/types";
 import { Project } from "../../models/project";
-import { ProjectLifecycle } from "../projects/ProjectLifecycle";
+import { ProjectCreationManager } from "../projects/ProjectCreationManager";
 import { DiffCache } from "./DiffCache";
 
 interface TemporaryProject {
@@ -13,7 +13,7 @@ interface TemporaryProject {
 }
 
 export class TemporaryProjectFactory {
-  private readonly lifecycle = new ProjectLifecycle({ git: false });
+  private readonly manager = new ProjectCreationManager({ git: false });
 
   constructor(private readonly cache = new DiffCache()) {}
 
@@ -30,7 +30,7 @@ export class TemporaryProjectFactory {
       await fs.rm(tempPathResult.data, { recursive: true, force: true });
     };
 
-    const generationResult = await this.lifecycle.generateFromTemplateSettings(
+    const generationResult = await this.manager.generateFromTemplateSettings(
       projectSettings,
       tempPathResult.data,
     );
@@ -61,7 +61,7 @@ export class TemporaryProjectFactory {
       await fs.rm(tempPathResult.data, { recursive: true, force: true });
     };
 
-    const generationResult = await this.lifecycle.generateFromExistingProject(
+    const generationResult = await this.manager.generateFromExistingProject(
       project,
       tempPathResult.data,
     );
